Make `kubeapiserver` component instantiation reuseable
