+++ pkg/operation/botanist/kubeapiserver_test.go
@@ -23,17 +23,12 @@ import (
 	"github.com/golang/mock/gomock"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-	. "github.com/onsi/gomega/gstruct"
 	gomegatypes "github.com/onsi/gomega/types"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/component-base/featuregate"
-	admissionapiv1 "k8s.io/pod-security-admission/admission/api/v1"
-	admissionapiv1alpha1 "k8s.io/pod-security-admission/admission/api/v1alpha1"
-	admissionapiv1beta1 "k8s.io/pod-security-admission/admission/api/v1beta1"
 	"k8s.io/utils/pointer"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	fakeclient "sigs.k8s.io/controller-runtime/pkg/client/fake"
@@ -81,6 +76,7 @@ var _ = Describe("KubeAPIServer", func() {
 		externalClusterDomain = "external.foo.bar.com"
 		podNetwork            *net.IPNet
 		serviceNetwork        *net.IPNet
+		seedVersion           = "1.22.0"
 		apiServerNetwork      = net.ParseIP("10.0.4.1")
 		podNetworkCIDR        = "10.0.1.0/24"
 		serviceNetworkCIDR    = "10.0.2.0/24"
@@ -94,7 +90,7 @@ var _ = Describe("KubeAPIServer", func() {
 
 		gardenClient = fakeclient.NewClientBuilder().WithScheme(kubernetes.GardenScheme).Build()
 		seedClient = fakeclient.NewClientBuilder().WithScheme(kubernetes.SeedScheme).Build()
-		seedClientSet = fake.NewClientSetBuilder().WithClient(seedClient).WithVersion("1.22.0").Build()
+		seedClientSet = fake.NewClientSetBuilder().WithClient(seedClient).WithVersion(seedVersion).Build()
 
 		var err error
 		_, podNetwork, err = net.ParseCIDR(podNetworkCIDR)
@@ -115,7 +111,9 @@ var _ = Describe("KubeAPIServer", func() {
 				SeedClientSet:  seedClientSet,
 				SecretsManager: sm,
 				Garden:         &garden.Garden{},
-				Seed:           &seedpkg.Seed{},
+				Seed: &seedpkg.Seed{
+					KubernetesVersion: semver.MustParse(seedVersion),
+				},
 				Shoot: &shootpkg.Shoot{
 					SeedNamespace: seedNamespace,
 					Components: &shootpkg.Components{
@@ -172,611 +170,6 @@ var _ = Describe("KubeAPIServer", func() {
 	})
 
 	Describe("#DefaultKubeAPIServer", func() {
-		It("should return an error because the apiserver-proxy-pod-webhook image cannot be found", func() {
-			botanist.ImageVector = imagevector.ImageVector{}
-
-			kubeAPIServer, err := botanist.DefaultKubeAPIServer(ctx)
-			Expect(kubeAPIServer).To(BeNil())
-			Expect(err).To(MatchError(ContainSubstring("could not find image \"apiserver-proxy-pod-webhook\"")))
-		})
-
-		It("should return an error because the kube-apiserver cannot be found", func() {
-			botanist.ImageVector = imagevector.ImageVector{{Name: "apiserver-proxy-pod-webhook"}}
-
-			kubeAPIServer, err := botanist.DefaultKubeAPIServer(ctx)
-			Expect(kubeAPIServer).To(BeNil())
-			Expect(err).To(MatchError(ContainSubstring("could not find image \"kube-apiserver\"")))
-		})
-
-		Describe("AnonymousAuthenticationEnabled", func() {
-			It("should set the field to false by default", func() {
-				kubeAPIServer, err := botanist.DefaultKubeAPIServer(ctx)
-				Expect(err).NotTo(HaveOccurred())
-				Expect(kubeAPIServer.GetValues().AnonymousAuthenticationEnabled).To(BeFalse())
-			})
-
-			It("should set the field to true if explicitly enabled", func() {
-				shootCopy := botanist.Shoot.GetInfo().DeepCopy()
-				shootCopy.Spec.Kubernetes.KubeAPIServer = &gardencorev1beta1.KubeAPIServerConfig{
-					EnableAnonymousAuthentication: pointer.Bool(true),
-				}
-
-				botanist.Shoot.SetInfo(shootCopy)
-
-				kubeAPIServer, err := botanist.DefaultKubeAPIServer(ctx)
-				Expect(err).NotTo(HaveOccurred())
-				Expect(kubeAPIServer.GetValues().AnonymousAuthenticationEnabled).To(BeTrue())
-			})
-		})
-
-		Describe("APIAudiences", func() {
-			It("should set the field to 'kubernetes' and 'gardener' by default", func() {
-				kubeAPIServer, err := botanist.DefaultKubeAPIServer(ctx)
-				Expect(err).NotTo(HaveOccurred())
-				Expect(kubeAPIServer.GetValues().APIAudiences).To(ConsistOf("kubernetes", "gardener"))
-			})
-
-			It("should set the field to the configured values", func() {
-				apiAudiences := []string{"foo", "bar"}
-
-				shootCopy := botanist.Shoot.GetInfo().DeepCopy()
-				shootCopy.Spec.Kubernetes.KubeAPIServer = &gardencorev1beta1.KubeAPIServerConfig{
-					APIAudiences: apiAudiences,
-				}
-				botanist.Shoot.SetInfo(shootCopy)
-
-				kubeAPIServer, err := botanist.DefaultKubeAPIServer(ctx)
-				Expect(err).NotTo(HaveOccurred())
-				Expect(kubeAPIServer.GetValues().APIAudiences).To(Equal(append(apiAudiences, "gardener")))
-			})
-
-			It("should not add gardener audience if already present", func() {
-				apiAudiences := []string{"foo", "bar", "gardener"}
-
-				shootCopy := botanist.Shoot.GetInfo().DeepCopy()
-				shootCopy.Spec.Kubernetes.KubeAPIServer = &gardencorev1beta1.KubeAPIServerConfig{
-					APIAudiences: apiAudiences,
-				}
-				botanist.Shoot.SetInfo(shootCopy)
-
-				kubeAPIServer, err := botanist.DefaultKubeAPIServer(ctx)
-				Expect(err).NotTo(HaveOccurred())
-				Expect(kubeAPIServer.GetValues().APIAudiences).To(Equal(apiAudiences))
-			})
-		})
-
-		Describe("AdmissionPlugins", func() {
-			var shootCopy *gardencorev1beta1.Shoot
-
-			BeforeEach(func() {
-				shootCopy = botanist.Shoot.GetInfo().DeepCopy()
-				shootCopy.Spec.Kubernetes = gardencorev1beta1.Kubernetes{
-					KubeAPIServer: &gardencorev1beta1.KubeAPIServerConfig{
-						AdmissionPlugins: []gardencorev1beta1.AdmissionPlugin{},
-					},
-					Version: "1.22.1",
-				}
-			})
-
-			DescribeTable("should have the expected admission plugins config",
-				func(configuredPlugins []gardencorev1beta1.AdmissionPlugin, expectedPlugins []kubeapiserver.AdmissionPluginConfig) {
-					shootCopy.Spec.Kubernetes.KubeAPIServer.AdmissionPlugins = configuredPlugins
-
-					botanist.Shoot.SetInfo(shootCopy)
-
-					kubeAPIServer, err := botanist.DefaultKubeAPIServer(ctx)
-					Expect(err).NotTo(HaveOccurred())
-					Expect(kubeAPIServer.GetValues().EnabledAdmissionPlugins).To(Equal(expectedPlugins))
-				},
-
-				Entry("only default plugins",
-					nil,
-					[]kubeapiserver.AdmissionPluginConfig{
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "Priority"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "NamespaceLifecycle"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "LimitRanger"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "PodSecurityPolicy"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "ServiceAccount"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "NodeRestriction"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "DefaultStorageClass"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "DefaultTolerationSeconds"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "ResourceQuota"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "StorageObjectInUseProtection"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "MutatingAdmissionWebhook"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "ValidatingAdmissionWebhook"}},
-					},
-				),
-				Entry("default plugins with overrides",
-					[]gardencorev1beta1.AdmissionPlugin{
-						{Name: "NamespaceLifecycle", Config: &runtime.RawExtension{Raw: []byte("namespace-lifecycle-config")}},
-					},
-					[]kubeapiserver.AdmissionPluginConfig{
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "Priority"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "NamespaceLifecycle", Config: &runtime.RawExtension{Raw: []byte("namespace-lifecycle-config")}}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "LimitRanger"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "PodSecurityPolicy"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "ServiceAccount"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "NodeRestriction"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "DefaultStorageClass"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "DefaultTolerationSeconds"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "ResourceQuota"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "StorageObjectInUseProtection"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "MutatingAdmissionWebhook"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "ValidatingAdmissionWebhook"}},
-					},
-				),
-				Entry("default plugins with overrides and other plugins",
-					[]gardencorev1beta1.AdmissionPlugin{
-						{Name: "NamespaceLifecycle", Config: &runtime.RawExtension{Raw: []byte("namespace-lifecycle-config")}},
-						{Name: "Foo"},
-						{Name: "Bar"},
-						{Name: "Baz", Config: &runtime.RawExtension{Raw: []byte("baz-config")}},
-					},
-					[]kubeapiserver.AdmissionPluginConfig{
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "Priority"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "NamespaceLifecycle", Config: &runtime.RawExtension{Raw: []byte("namespace-lifecycle-config")}}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "LimitRanger"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "PodSecurityPolicy"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "ServiceAccount"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "NodeRestriction"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "DefaultStorageClass"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "DefaultTolerationSeconds"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "ResourceQuota"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "StorageObjectInUseProtection"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "MutatingAdmissionWebhook"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "ValidatingAdmissionWebhook"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "Foo"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "Bar"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "Baz", Config: &runtime.RawExtension{Raw: []byte("baz-config")}}},
-					},
-				),
-				Entry("default plugins with overrides and skipping configured plugins if disabled",
-					[]gardencorev1beta1.AdmissionPlugin{
-						{Name: "NamespaceLifecycle", Config: &runtime.RawExtension{Raw: []byte("namespace-lifecycle-config")}},
-						{Name: "Foo"},
-						{Name: "Bar", Disabled: pointer.Bool(true)},
-						{Name: "Baz", Config: &runtime.RawExtension{Raw: []byte("baz-config")}, Disabled: pointer.Bool(true)},
-					},
-					[]kubeapiserver.AdmissionPluginConfig{
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "Priority"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "NamespaceLifecycle", Config: &runtime.RawExtension{Raw: []byte("namespace-lifecycle-config")}}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "LimitRanger"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "PodSecurityPolicy"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "ServiceAccount"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "NodeRestriction"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "DefaultStorageClass"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "DefaultTolerationSeconds"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "ResourceQuota"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "StorageObjectInUseProtection"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "MutatingAdmissionWebhook"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "ValidatingAdmissionWebhook"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "Foo"}},
-					},
-				),
-				Entry("default plugins with overrides and skipping default plugins if disabled",
-					[]gardencorev1beta1.AdmissionPlugin{
-						{Name: "NamespaceLifecycle", Config: &runtime.RawExtension{Raw: []byte("namespace-lifecycle-config")}},
-						{Name: "PodSecurityPolicy", Disabled: pointer.Bool(true)},
-						{Name: "ResourceQuota", Disabled: pointer.Bool(true)},
-						{Name: "Foo"},
-						{Name: "Bar"},
-						{Name: "Baz", Config: &runtime.RawExtension{Raw: []byte("baz-config")}},
-						{Name: "ServiceAccount", Disabled: pointer.Bool(false)},
-					},
-					[]kubeapiserver.AdmissionPluginConfig{
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "Priority"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "NamespaceLifecycle", Config: &runtime.RawExtension{Raw: []byte("namespace-lifecycle-config")}}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "LimitRanger"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "ServiceAccount", Disabled: pointer.Bool(false)}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "NodeRestriction"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "DefaultStorageClass"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "DefaultTolerationSeconds"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "StorageObjectInUseProtection"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "MutatingAdmissionWebhook"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "ValidatingAdmissionWebhook"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "Foo"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "Bar"}},
-						{AdmissionPlugin: gardencorev1beta1.AdmissionPlugin{Name: "Baz", Config: &runtime.RawExtension{Raw: []byte("baz-config")}}},
-					},
-				),
-			)
-
-			Context("should have the expected disabled admission plugins", func() {
-				var expectedDisabledPlugins []gardencorev1beta1.AdmissionPlugin
-
-				AfterEach(func() {
-					botanist.Shoot.SetInfo(shootCopy)
-					kubeAPIServer, err := botanist.DefaultKubeAPIServer(ctx)
-					Expect(err).NotTo(HaveOccurred())
-					Expect(kubeAPIServer.GetValues().DisabledAdmissionPlugins).To(Equal(expectedDisabledPlugins))
-				})
-
-				It("should return the correct list of disabled admission plugins", func() {
-					shootCopy.Spec.Kubernetes.KubeAPIServer.AdmissionPlugins = []gardencorev1beta1.AdmissionPlugin{
-						{Name: "Priority"},
-						{Name: "NamespaceLifecycle", Config: &runtime.RawExtension{Raw: []byte("namespace-lifecycle-config")}},
-						{Name: "LimitRanger"},
-						{Name: "PodSecurityPolicy", Disabled: pointer.Bool(true)},
-						{Name: "ServiceAccount"},
-						{Name: "NodeRestriction"},
-						{Name: "DefaultStorageClass"},
-						{Name: "DefaultTolerationSeconds", Disabled: pointer.Bool(true)},
-						{Name: "ResourceQuota"},
-					}
-
-					expectedDisabledPlugins = []gardencorev1beta1.AdmissionPlugin{
-						{Name: "PodSecurityPolicy", Disabled: pointer.Bool(true)},
-						{Name: "DefaultTolerationSeconds", Disabled: pointer.Bool(true)},
-					}
-				})
-
-				It("should return the correct list of disabled admission plugins", func() {
-					shootCopy.Spec.Kubernetes.KubeAPIServer.AdmissionPlugins = []gardencorev1beta1.AdmissionPlugin{
-						{Name: "Priority"},
-						{Name: "NamespaceLifecycle", Config: &runtime.RawExtension{Raw: []byte("namespace-lifecycle-config")}, Disabled: pointer.Bool(true)},
-						{Name: "LimitRanger"},
-						{Name: "PodSecurityPolicy"},
-						{Name: "ServiceAccount"},
-						{Name: "NodeRestriction"},
-						{Name: "DefaultStorageClass", Disabled: pointer.Bool(true)},
-						{Name: "DefaultTolerationSeconds"},
-						{Name: "ResourceQuota"},
-						{Name: "foo", Config: &runtime.RawExtension{Raw: []byte("foo-config")}, Disabled: pointer.Bool(true)},
-					}
-
-					expectedDisabledPlugins = []gardencorev1beta1.AdmissionPlugin{
-						{Name: "NamespaceLifecycle", Config: &runtime.RawExtension{Raw: []byte("namespace-lifecycle-config")}, Disabled: pointer.Bool(true)},
-						{Name: "DefaultStorageClass", Disabled: pointer.Bool(true)},
-						{Name: "foo", Config: &runtime.RawExtension{Raw: []byte("foo-config")}, Disabled: pointer.Bool(true)},
-					}
-				})
-			})
-
-			Describe("PodSecurity Admission Plugin", func() {
-				var (
-					configData    *runtime.RawExtension
-					err           error
-					kubeAPIServer kubeapiserver.Interface
-				)
-				JustBeforeEach(func() {
-					configData = nil
-					botanist.Shoot.SetInfo(shootCopy)
-					kubeAPIServer, err = botanist.DefaultKubeAPIServer(ctx)
-				})
-
-				Context("When the config is nil", func() {
-					It("should do nothing", func() {
-						Expect(err).NotTo(HaveOccurred())
-
-						admissionPlugins := kubeAPIServer.GetValues().EnabledAdmissionPlugins
-						for _, plugin := range admissionPlugins {
-							if plugin.Name == "PodSecurity" {
-								configData = plugin.Config
-							}
-						}
-
-						Expect(configData).To(BeNil())
-					})
-				})
-
-				Context("When the config is not nil", func() {
-					var (
-						err    error
-						ok     bool
-						config runtime.Object
-					)
-
-					JustBeforeEach(func() {
-						admissionPlugins := kubeAPIServer.GetValues().EnabledAdmissionPlugins
-						for _, plugin := range admissionPlugins {
-							if plugin.Name == "PodSecurity" {
-								configData = plugin.Config
-							}
-						}
-
-						Expect(configData).NotTo(BeNil())
-
-						config, err = runtime.Decode(codec, configData.Raw)
-						Expect(err).NotTo(HaveOccurred())
-					})
-
-					Context("PodSecurity admission config is v1", func() {
-						BeforeEach(func() {
-							shootCopy.Spec.Kubernetes.KubeAPIServer.AdmissionPlugins = []gardencorev1beta1.AdmissionPlugin{
-								{
-									Name: "PodSecurity",
-									Config: &runtime.RawExtension{Raw: []byte(`apiVersion: pod-security.admission.config.k8s.io/v1
-kind: PodSecurityConfiguration
-defaults:
-  enforce: "privileged"
-  enforce-version: "latest"
-  audit-version: "latest"
-  warn: "baseline"
-  warn-version: "v1.23"
-exemptions:
-  usernames: ["admin"]
-  runtimeClasses: ["random"]
-  namespaces: ["random"]
-`),
-									},
-								},
-							}
-						})
-
-						It("should add kube-system to exempted namespaces and not touch other fields", func() {
-							var admConfig *admissionapiv1.PodSecurityConfiguration
-
-							admConfig, ok = config.(*admissionapiv1.PodSecurityConfiguration)
-							Expect(ok).To(BeTrue())
-
-							Expect(admConfig.Defaults).To(MatchFields(IgnoreExtras, Fields{
-								"Enforce": Equal("privileged"),
-								// This field is defaulted by kubernetes
-								"Audit":       Equal("privileged"),
-								"Warn":        Equal("baseline"),
-								"WarnVersion": Equal("v1.23"),
-							}))
-							Expect(admConfig.Exemptions.Usernames).To(ContainElement("admin"))
-							Expect(admConfig.Exemptions.Namespaces).To(ContainElements("kube-system", "random"))
-						})
-					})
-
-					Context("PodSecurity admission config is v1beta1", func() {
-						BeforeEach(func() {
-							shootCopy.Spec.Kubernetes.KubeAPIServer.AdmissionPlugins = []gardencorev1beta1.AdmissionPlugin{
-								{
-									Name: "PodSecurity",
-									Config: &runtime.RawExtension{Raw: []byte(`apiVersion: pod-security.admission.config.k8s.io/v1beta1
-kind: PodSecurityConfiguration
-defaults:
-  enforce: "privileged"
-  enforce-version: "latest"
-  audit-version: "latest"
-  warn: "baseline"
-  warn-version: "v1.23"
-exemptions:
-  usernames: ["admin"]
-  runtimeClasses: ["random"]
-  namespaces: ["random"]
-`),
-									},
-								},
-							}
-						})
-
-						It("should add kube-system to exempted namespaces and not touch other fields", func() {
-							var admConfig *admissionapiv1beta1.PodSecurityConfiguration
-
-							admConfig, ok = config.(*admissionapiv1beta1.PodSecurityConfiguration)
-							Expect(ok).To(BeTrue())
-
-							Expect(admConfig.Defaults).To(MatchFields(IgnoreExtras, Fields{
-								"Enforce": Equal("privileged"),
-								// This field is defaulted by kubernetes
-								"Audit":       Equal("privileged"),
-								"Warn":        Equal("baseline"),
-								"WarnVersion": Equal("v1.23"),
-							}))
-							Expect(admConfig.Exemptions.Usernames).To(ContainElement("admin"))
-							Expect(admConfig.Exemptions.Namespaces).To(ContainElements("kube-system", "random"))
-						})
-					})
-
-					Context("PodSecurity admission config is v1alpha1", func() {
-						BeforeEach(func() {
-							shootCopy.Spec.Kubernetes.KubeAPIServer.AdmissionPlugins = []gardencorev1beta1.AdmissionPlugin{
-								{
-									Name: "PodSecurity",
-									Config: &runtime.RawExtension{Raw: []byte(`apiVersion: pod-security.admission.config.k8s.io/v1alpha1
-kind: PodSecurityConfiguration
-defaults:
-  enforce: "privileged"
-  enforce-version: "latest"
-  audit-version: "latest"
-  warn: "baseline"
-  warn-version: "v1.22"
-exemptions:
-  usernames: ["admin"]
-  runtimeClasses: ["random"]
-  namespaces: ["random"]
-`),
-									},
-								},
-							}
-						})
-
-						It("should add kube-system to exempted namespaces and not touch other fields", func() {
-							var admConfig *admissionapiv1alpha1.PodSecurityConfiguration
-
-							admConfig, ok = config.(*admissionapiv1alpha1.PodSecurityConfiguration)
-							Expect(ok).To(BeTrue())
-
-							Expect(admConfig.Defaults).To(MatchFields(IgnoreExtras, Fields{
-								"Enforce": Equal("privileged"),
-								// This field is defaulted by kubernetes
-								"Audit":       Equal("privileged"),
-								"Warn":        Equal("baseline"),
-								"WarnVersion": Equal("v1.22"),
-							}))
-							Expect(admConfig.Exemptions.Usernames).To(ContainElement("admin"))
-							Expect(admConfig.Exemptions.Namespaces).To(ContainElements("kube-system", "random"))
-						})
-					})
-				})
-
-				Context("PodSecurity admission config is neither v1alpha1 nor v1beta1 nor v1", func() {
-					BeforeEach(func() {
-						shootCopy.Spec.Kubernetes.KubeAPIServer.AdmissionPlugins = []gardencorev1beta1.AdmissionPlugin{
-							{
-								Name: "PodSecurity",
-								Config: &runtime.RawExtension{Raw: []byte(`apiVersion: pod-security.admission.config.k8s.io/foo
-kind: PodSecurityConfiguration-bar
-defaults:
-  enforce: "privileged"
-  enforce-version: "latest"
-exemptions:
-  usernames: ["admin"]
-`),
-								},
-							},
-						}
-					})
-
-					It("should throw an error", func() {
-						Expect(kubeAPIServer).To(BeNil())
-
-						Expect(err).To(BeNotRegisteredError())
-					})
-				})
-			})
-		})
-
-		Describe("AuditConfig", func() {
-			var (
-				policy               = "some-policy"
-				auditPolicyConfigMap *corev1.ConfigMap
-			)
-
-			BeforeEach(func() {
-				auditPolicyConfigMap = &corev1.ConfigMap{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "my-audit-policy",
-						Namespace: projectNamespace,
-					},
-					Data: map[string]string{"policy": policy},
-				}
-			})
-
-			DescribeTable("should have the expected audit config",
-				func(prepTest func(), expectedConfig *kubeapiserver.AuditConfig, errMatcher gomegatypes.GomegaMatcher) {
-					if prepTest != nil {
-						prepTest()
-					}
-
-					kubeAPIServer, err := botanist.DefaultKubeAPIServer(ctx)
-					Expect(err).To(errMatcher)
-					if kubeAPIServer != nil {
-						Expect(kubeAPIServer.GetValues().Audit).To(Equal(expectedConfig))
-					}
-				},
-
-				Entry("KubeAPIServerConfig is nil",
-					nil,
-					nil,
-					Not(HaveOccurred()),
-				),
-				Entry("AuditConfig is nil",
-					func() {
-						shootCopy := botanist.Shoot.GetInfo().DeepCopy()
-						shootCopy.Spec.Kubernetes.KubeAPIServer = &gardencorev1beta1.KubeAPIServerConfig{}
-						botanist.Shoot.SetInfo(shootCopy)
-					},
-					nil,
-					Not(HaveOccurred()),
-				),
-				Entry("AuditPolicy is nil",
-					func() {
-						shootCopy := botanist.Shoot.GetInfo().DeepCopy()
-						shootCopy.Spec.Kubernetes.KubeAPIServer = &gardencorev1beta1.KubeAPIServerConfig{
-							AuditConfig: &gardencorev1beta1.AuditConfig{},
-						}
-						botanist.Shoot.SetInfo(shootCopy)
-					},
-					nil,
-					Not(HaveOccurred()),
-				),
-				Entry("ConfigMapRef is nil",
-					func() {
-						shootCopy := botanist.Shoot.GetInfo().DeepCopy()
-						shootCopy.Spec.Kubernetes.KubeAPIServer = &gardencorev1beta1.KubeAPIServerConfig{
-							AuditConfig: &gardencorev1beta1.AuditConfig{
-								AuditPolicy: &gardencorev1beta1.AuditPolicy{},
-							},
-						}
-						botanist.Shoot.SetInfo(shootCopy)
-					},
-					nil,
-					Not(HaveOccurred()),
-				),
-				Entry("ConfigMapRef is provided but configmap is missing",
-					func() {
-						shootCopy := botanist.Shoot.GetInfo().DeepCopy()
-						shootCopy.Spec.Kubernetes.KubeAPIServer = &gardencorev1beta1.KubeAPIServerConfig{
-							AuditConfig: &gardencorev1beta1.AuditConfig{
-								AuditPolicy: &gardencorev1beta1.AuditPolicy{
-									ConfigMapRef: &corev1.ObjectReference{
-										Name: auditPolicyConfigMap.Name,
-									},
-								},
-							},
-						}
-						botanist.Shoot.SetInfo(shootCopy)
-					},
-					nil,
-					MatchError(ContainSubstring("not found")),
-				),
-				Entry("ConfigMapRef is provided but configmap is missing while shoot has a deletion timestamp",
-					func() {
-						shootCopy := botanist.Shoot.GetInfo().DeepCopy()
-						shootCopy.DeletionTimestamp = &metav1.Time{}
-						shootCopy.Spec.Kubernetes.KubeAPIServer = &gardencorev1beta1.KubeAPIServerConfig{
-							AuditConfig: &gardencorev1beta1.AuditConfig{
-								AuditPolicy: &gardencorev1beta1.AuditPolicy{
-									ConfigMapRef: &corev1.ObjectReference{
-										Name: auditPolicyConfigMap.Name,
-									},
-								},
-							},
-						}
-						botanist.Shoot.SetInfo(shootCopy)
-					},
-					&kubeapiserver.AuditConfig{},
-					Not(HaveOccurred()),
-				),
-				Entry("ConfigMapRef is provided but configmap does not have correct data field",
-					func() {
-						auditPolicyConfigMap.Data = nil
-						Expect(gardenClient.Create(ctx, auditPolicyConfigMap)).To(Succeed())
-
-						shootCopy := botanist.Shoot.GetInfo().DeepCopy()
-						shootCopy.Spec.Kubernetes.KubeAPIServer = &gardencorev1beta1.KubeAPIServerConfig{
-							AuditConfig: &gardencorev1beta1.AuditConfig{
-								AuditPolicy: &gardencorev1beta1.AuditPolicy{
-									ConfigMapRef: &corev1.ObjectReference{
-										Name: auditPolicyConfigMap.Name,
-									},
-								},
-							},
-						}
-						botanist.Shoot.SetInfo(shootCopy)
-					},
-					nil,
-					MatchError(ContainSubstring("missing '.data.policy' in audit policy configmap")),
-				),
-				Entry("ConfigMapRef is provided and configmap is compliant",
-					func() {
-						Expect(gardenClient.Create(ctx, auditPolicyConfigMap)).To(Succeed())
-
-						shootCopy := botanist.Shoot.GetInfo().DeepCopy()
-						shootCopy.Spec.Kubernetes.KubeAPIServer = &gardencorev1beta1.KubeAPIServerConfig{
-							AuditConfig: &gardencorev1beta1.AuditConfig{
-								AuditPolicy: &gardencorev1beta1.AuditPolicy{
-									ConfigMapRef: &corev1.ObjectReference{
-										Name: auditPolicyConfigMap.Name,
-									},
-								},
-							},
-						}
-						botanist.Shoot.SetInfo(shootCopy)
-					},
-					&kubeapiserver.AuditConfig{
-						Policy: &policy,
-					},
-					Not(HaveOccurred()),
-				),
-			)
-		})
-
 		Describe("AutoscalingConfig", func() {
 			DescribeTable("should have the expected autoscaling config",
 				func(prepTest func(), featureGate *featuregate.Feature, value *bool, expectedConfig kubeapiserver.AutoscalingConfig) {
@@ -941,219 +334,6 @@ exemptions:
 				),
 			)
 		})
-
-		Describe("DefaultNotReadyTolerationSeconds and DefaultUnreachableTolerationSeconds", func() {
-			It("should not set the fields", func() {
-				kubeAPIServer, err := botanist.DefaultKubeAPIServer(ctx)
-				Expect(err).NotTo(HaveOccurred())
-				Expect(kubeAPIServer.GetValues().DefaultNotReadyTolerationSeconds).To(BeNil())
-				Expect(kubeAPIServer.GetValues().DefaultUnreachableTolerationSeconds).To(BeNil())
-			})
-
-			It("should set the fields to the configured values", func() {
-				shootCopy := botanist.Shoot.GetInfo().DeepCopy()
-				shootCopy.Spec.Kubernetes.KubeAPIServer = &gardencorev1beta1.KubeAPIServerConfig{
-					DefaultNotReadyTolerationSeconds:    pointer.Int64(120),
-					DefaultUnreachableTolerationSeconds: pointer.Int64(130),
-				}
-				botanist.Shoot.SetInfo(shootCopy)
-
-				kubeAPIServer, err := botanist.DefaultKubeAPIServer(ctx)
-				Expect(err).NotTo(HaveOccurred())
-				Expect(kubeAPIServer.GetValues().DefaultNotReadyTolerationSeconds).To(PointTo(Equal(int64(120))))
-				Expect(kubeAPIServer.GetValues().DefaultUnreachableTolerationSeconds).To(PointTo(Equal(int64(130))))
-			})
-		})
-
-		Describe("EventTTL", func() {
-			It("should not set the event ttl field", func() {
-				kubeAPIServer, err := botanist.DefaultKubeAPIServer(ctx)
-				Expect(err).NotTo(HaveOccurred())
-				Expect(kubeAPIServer.GetValues().EventTTL).To(BeNil())
-			})
-
-			It("should set the field to the configured values", func() {
-				eventTTL := &metav1.Duration{Duration: 2 * time.Hour}
-
-				shootCopy := botanist.Shoot.GetInfo().DeepCopy()
-				shootCopy.Spec.Kubernetes.KubeAPIServer = &gardencorev1beta1.KubeAPIServerConfig{
-					EventTTL: eventTTL,
-				}
-				botanist.Shoot.SetInfo(shootCopy)
-
-				kubeAPIServer, err := botanist.DefaultKubeAPIServer(ctx)
-				Expect(err).NotTo(HaveOccurred())
-				Expect(kubeAPIServer.GetValues().EventTTL).To(Equal(eventTTL))
-			})
-		})
-
-		Describe("FeatureGates", func() {
-			It("should set the field to nil by default", func() {
-				kubeAPIServer, err := botanist.DefaultKubeAPIServer(ctx)
-				Expect(err).NotTo(HaveOccurred())
-				Expect(kubeAPIServer.GetValues().FeatureGates).To(BeNil())
-			})
-
-			It("should set the field to the configured values", func() {
-				featureGates := map[string]bool{"foo": true, "bar": false}
-
-				shootCopy := botanist.Shoot.GetInfo().DeepCopy()
-				shootCopy.Spec.Kubernetes.KubeAPIServer = &gardencorev1beta1.KubeAPIServerConfig{
-					KubernetesConfig: gardencorev1beta1.KubernetesConfig{
-						FeatureGates: featureGates,
-					},
-				}
-				botanist.Shoot.SetInfo(shootCopy)
-
-				kubeAPIServer, err := botanist.DefaultKubeAPIServer(ctx)
-				Expect(err).NotTo(HaveOccurred())
-				Expect(kubeAPIServer.GetValues().FeatureGates).To(Equal(featureGates))
-			})
-		})
-
-		Describe("OIDCConfig", func() {
-			DescribeTable("should have the expected OIDC config",
-				func(prepTest func(), expectedConfig *gardencorev1beta1.OIDCConfig) {
-					if prepTest != nil {
-						prepTest()
-					}
-
-					kubeAPIServer, err := botanist.DefaultKubeAPIServer(ctx)
-					Expect(err).NotTo(HaveOccurred())
-					Expect(kubeAPIServer.GetValues().OIDC).To(Equal(expectedConfig))
-				},
-
-				Entry("KubeAPIServerConfig is nil",
-					nil,
-					nil,
-				),
-				Entry("OIDCConfig is nil",
-					func() {
-						shootCopy := botanist.Shoot.GetInfo().DeepCopy()
-						shootCopy.Spec.Kubernetes.KubeAPIServer = &gardencorev1beta1.KubeAPIServerConfig{}
-						botanist.Shoot.SetInfo(shootCopy)
-					},
-					nil,
-				),
-				Entry("OIDCConfig is not nil",
-					func() {
-						shootCopy := botanist.Shoot.GetInfo().DeepCopy()
-						shootCopy.Spec.Kubernetes.KubeAPIServer = &gardencorev1beta1.KubeAPIServerConfig{
-							OIDCConfig: &gardencorev1beta1.OIDCConfig{},
-						}
-						botanist.Shoot.SetInfo(shootCopy)
-					},
-					&gardencorev1beta1.OIDCConfig{},
-				),
-			)
-		})
-
-		Describe("Requests", func() {
-			It("should set the field to nil by default", func() {
-				kubeAPIServer, err := botanist.DefaultKubeAPIServer(ctx)
-				Expect(err).NotTo(HaveOccurred())
-				Expect(kubeAPIServer.GetValues().Requests).To(BeNil())
-			})
-
-			It("should set the field to the configured values", func() {
-				requests := &gardencorev1beta1.KubeAPIServerRequests{
-					MaxMutatingInflight:    pointer.Int32(1),
-					MaxNonMutatingInflight: pointer.Int32(2),
-				}
-
-				shootCopy := botanist.Shoot.GetInfo().DeepCopy()
-				shootCopy.Spec.Kubernetes.KubeAPIServer = &gardencorev1beta1.KubeAPIServerConfig{
-					Requests: requests,
-				}
-				botanist.Shoot.SetInfo(shootCopy)
-
-				kubeAPIServer, err := botanist.DefaultKubeAPIServer(ctx)
-				Expect(err).NotTo(HaveOccurred())
-				Expect(kubeAPIServer.GetValues().Requests).To(Equal(requests))
-			})
-		})
-
-		Describe("RuntimeConfig", func() {
-			It("should set the field to nil by default", func() {
-				kubeAPIServer, err := botanist.DefaultKubeAPIServer(ctx)
-				Expect(err).NotTo(HaveOccurred())
-				Expect(kubeAPIServer.GetValues().RuntimeConfig).To(BeNil())
-			})
-
-			It("should set the field to the configured values", func() {
-				runtimeConfig := map[string]bool{"foo": true, "bar": false}
-
-				shootCopy := botanist.Shoot.GetInfo().DeepCopy()
-				shootCopy.Spec.Kubernetes.KubeAPIServer = &gardencorev1beta1.KubeAPIServerConfig{
-					RuntimeConfig: runtimeConfig,
-				}
-				botanist.Shoot.SetInfo(shootCopy)
-
-				kubeAPIServer, err := botanist.DefaultKubeAPIServer(ctx)
-				Expect(err).NotTo(HaveOccurred())
-				Expect(kubeAPIServer.GetValues().RuntimeConfig).To(Equal(runtimeConfig))
-			})
-		})
-
-		Describe("VPNConfig", func() {
-			DescribeTable("should have the expected VPN config",
-				func(prepTest func(), expectedConfig kubeapiserver.VPNConfig) {
-					if prepTest != nil {
-						prepTest()
-					}
-
-					kubeAPIServer, err := botanist.DefaultKubeAPIServer(ctx)
-					Expect(err).NotTo(HaveOccurred())
-					Expect(kubeAPIServer.GetValues().VPN).To(Equal(expectedConfig))
-				},
-
-				Entry("ReversedVPN enabled",
-					nil,
-					kubeapiserver.VPNConfig{
-						Enabled:         true,
-						PodNetworkCIDR:  podNetworkCIDR,
-						NodeNetworkCIDR: &nodeNetworkCIDR,
-					},
-				),
-				Entry("no node network",
-					func() {
-						shootCopy := botanist.Shoot.GetInfo().DeepCopy()
-						shootCopy.Spec.Networking.Nodes = nil
-						botanist.Shoot.SetInfo(shootCopy)
-					},
-					kubeapiserver.VPNConfig{
-						Enabled:        true,
-						PodNetworkCIDR: podNetworkCIDR,
-					},
-				),
-			)
-		})
-
-		Describe("WatchCacheSizes", func() {
-			It("should set the field to nil by default", func() {
-				kubeAPIServer, err := botanist.DefaultKubeAPIServer(ctx)
-				Expect(err).NotTo(HaveOccurred())
-				Expect(kubeAPIServer.GetValues().WatchCacheSizes).To(BeNil())
-			})
-
-			It("should set the field to the configured values", func() {
-				watchCacheSizes := &gardencorev1beta1.WatchCacheSizes{
-					Default:   pointer.Int32(1),
-					Resources: []gardencorev1beta1.ResourceWatchCacheSize{{Resource: "foo"}},
-				}
-
-				shootCopy := botanist.Shoot.GetInfo().DeepCopy()
-				shootCopy.Spec.Kubernetes.KubeAPIServer = &gardencorev1beta1.KubeAPIServerConfig{
-					WatchCacheSizes: watchCacheSizes,
-				}
-				botanist.Shoot.SetInfo(shootCopy)
-
-				kubeAPIServer, err := botanist.DefaultKubeAPIServer(ctx)
-				Expect(err).NotTo(HaveOccurred())
-				Expect(kubeAPIServer.GetValues().WatchCacheSizes).To(Equal(watchCacheSizes))
-			})
-		})
-
 	})
 
 	DescribeTable("#resourcesRequirementsForKubeAPIServer",
